TS: fix start, and remove unused code
