+++ @reface/types/mod.ts
@@ -1,6 +1,3 @@
-export * from "./base.types.ts";
-export * from "./html.types.ts";
-export * from "./component.types.ts";
-export * from "./template.types.ts";
-export * from "./render.types.ts";
 export * from "./composer.types.ts";
+export * from "./render.types.ts";
+export * from "../template/types.ts";
