+++ @reface/types/render.types.ts
@@ -1,6 +1,4 @@
-import type { IRefaceRenderManager } from "./composer.types.ts";
-import type { ClassValue, HTMLAttributes, StyleValue } from "./html.types.ts";
-import type { IRefaceTemplate } from "./template.types.ts";
+import type { Template } from "@reface/template";
 
 export type RenderPhaseBase =
   | "render"
@@ -14,14 +12,21 @@ export type RenderPhaseBase =
 export type RenderPhase = `${RenderPhaseBase}:start` | `${RenderPhaseBase}:end`;
 
 export type RenderContext = {
-  template?: IRefaceTemplate;
+  template?: Template;
   html?: string;
-  attributes?: HTMLAttributes;
-  class?: ClassValue;
-  style?: StyleValue;
-  manager: IRefaceRenderManager;
+  attributes?: Record<string, any>;
+  class?: string[] | Record<string, boolean>;
+  style?: Record<string, string | number>;
+  manager: RenderManager;
 };
 
 export type RenderHandler = (
   context: RenderContext,
-) => void | IRefaceTemplate | string;
+) => void | Template | string;
+
+export interface RenderManager {
+  on(event: RenderPhase, handler: RenderHandler): void;
+  off(event: RenderPhase, handler: RenderHandler): void;
+  emit(event: RenderPhase, context: RenderContext): void;
+  render(template: Template): string;
+}
