+++ @reface/utils/mod.ts
@@ -1,2 +1,1 @@
 export * from "./escape.ts";
-export * from "./attributes.ts";
